forum

+++ View/subjectsView.php
@@ -1,43 +1,27 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Forum : Accueil</title>
-    <script src="https://kit.fontawesome.com/351e9300a0.js" crossorigin="anonymous"></script>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/assets/css/style.css">
-</head>
-<body>
-
-<div id="wrap">
-    <header>
-        <div id="menu" class="flexCenter flexRow">
-            <a href="../index.php">Accueil</a>
-            <a href="connectionView.php">Connexion</a>
-            <a href="registrationView.php">Inscription</a>
-            <a href="#">Compte</a>
-        </div>
-    </header>
-
-
     <main>
-        <h1 class="center">Cuisines</h1>
-        <div class="height">
-            <a href="Update/updateCategorieView.php" class="button buttonAbsolute1"><i class="fas fa-edit"></i></a>
-        </div>
-        <div class="height">
-            <a href="#" class="button buttonAbsolute1"><i class="fas fa-archive"></i></a>
-        </div>
-        <div class="height">
-            <a href="Delete/deleteCategorieView.php" class="button buttonAbsolute1"><i class="fas fa-trash-alt"></i></a>
-        </div>
-        <div class="flexCenter">
-            <img class="imageCategorie" src="https://images.unsplash.com/photo-1617196034564-65baf56380ab?ixid=MnwxMjA3fDB8MHxzZWFyY2h8NHx8Y3VzaW5lfGVufDB8fDB8fA%3D%3D&ixlib=rb-1.2.1&auto=format&fit=crop&w=500&q=60">
-        </div>
+        <?php
+        if (isset($var['categorie'])) {
+            foreach ($var['categorie'] as $categorie) {?>
+                <h1 class="center"><?=$categorie->getTitle()?></h1>
+                <div class="height">
+                    <a href="../index.php?controller=categories&action=update&id=<?=$categorie->getId()?>" class="button buttonAbsolute1"><i class="fas fa-edit"></i></a>
+                </div>
+                <div class="height">
+                    <a href="#" class="button buttonAbsolute1"><i class="fas fa-archive"></i></a>
+                </div>
+                <div class="height">
+                    <a href="Delete/deleteCategorieView.php" class="button buttonAbsolute1"><i class="fas fa-trash-alt"></i></a>
+                </div>
+                <div class="flexCenter">
+                    <img class="imageCategorie" src="<?=$categorie->getPicture()?>" alt="<?=$categorie->getTitle()?>">
+                </div>
 
-        <div class="height">
-            <a href="Create/createSubjectView.php" class="button buttonAbsolute1"><i class="fas fa-plus"></i> Ajouter un sujet</a>
-        </div>
+                <div class="height">
+                    <a href="Create/createSubjectView.php" class="button buttonAbsolute1"><i class="fas fa-plus"></i> Ajouter un sujet</a>
+                </div>
+            <?php
+            }
+        }?>
 
         <div id="containerSubjects" class="flexCenter flexRow wrap">
             <div class="border flexRow">
@@ -51,48 +35,23 @@
                     <p>Dates</p>
                 </div>
             </div>
-            <a href="subjectView.php" class="border1 flexRow">
-                <div class="width20 flexCenter">
-                    <p>Tartiflette</p>
-                </div>
-                <div class="width60 flexCenter wrap">
-                    <p>blablablablablbalblablablablablablabalablabalbalabalabal</p>
-                </div>
-                <div class="width20 flexCenter">
-                    <p>00/00/0000</p>
-                </div>
-            </a>
-            <a href="#" class="border1 flexRow">
-                <div class="width20 flexCenter">
-                    <p>Tartiflette</p>
-                </div>
-                <div class="width60 flexCenter wrap">
-                    <p>blablablablablbalblablablablablablabalablabalbalabalabal</p>
-                </div>
-                <div class="width20 flexCenter">
-                    <p>00/00/0000</p>
-                </div>
-            </a>
-            <a href="#" class="border1 flexRow">
-                <div class="width20 flexCenter">
-                    <p>Tartiflette</p>
-                </div>
-                <div class="width60 flexCenter wrap">
-                    <p>blablablablablbalblablablablablablabalablabalbalabalabal</p>
-                </div>
-                <div class="width20 flexCenter">
-                    <p>00/00/0000</p>
-                </div>
-            </a>
+            <?php
+            if (isset($var['subjects'])) {
+                foreach ($var['subjects'] as $subject) {?>
+                    <a href="#" class="border1 flexRow">
+                        <div class="width20 flexCenter">
+                            <p><?=$subject->getTitle()?></p>
+                        </div>
+                        <div class="width60 flexCenter wrap">
+                            <p><?=$subject->getDescription() ?></p>
+                        </div>
+                        <div class="width20 flexCenter">
+                            <p><?=$subject->getDate()?></p>
+                        </div>
+                    </a>
+                    <?php
+                }
+            }
+            ?>
         </div>
     </main>
-
-    <footer class="flexCenter">
-        <a href="">Mentions légales</a>
-        <a href="">Contact</a>
-    </footer>
-</div>
-<script src="/assets/js/app.js"></script>
-
-</body>
-</html>
